third_party/WebKit/LayoutTests/editing/selection/readonly-disabled-text-selection-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS window.getSelection().toString() is "NormalInput"
PASS window.getSelection().toString() is "ReadonlyInput"
PASS window.getSelection().toString() is "DisabledInput"
PASS window.getSelection().toString() is "ReadonlyDisabledInput"
PASS window.getSelection().toString() is "NormalTextarea"
PASS window.getSelection().toString() is "ReadonlyTextarea"
PASS window.getSelection().toString() is "DisabledTextarea"
PASS window.getSelection().toString() is "ReadonlyDisabledTextarea"
       
